Moved policies declarations to individual files
