+++ utils/auth/policies/index.ts
@@ -1,0 +1,7 @@
+import allowRedPolicy from './allowRedPolicy'
+
+const Policies = {
+  allowRed: allowRedPolicy,
+}
+
+export default Policies
